S2L set_selenium_speed and webdrivermonkeypatches re-factored
By default, S2L did always monkey patch the WebDriver execute method.
But the monkey patch was actually needed actually only needed when the
S2L set_selenium_speed method was called. Re-factored the code, to
perform the monkey patching only when the S2L set_selenium_speed
method is called.
